mockVision/pyproject.toml
[project]
name = "mockvision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pyautogui>=0.9.54",
    "uvicorn>=0.35.0",
]
